:ok_hand: IMPROVE: Re-organize tracker module
